# virtio-fs: New case to test virtio-fs xfstest