--- a/qemu/tests/cfg/virtio_fs_share_data.cfg
+++ b/qemu/tests/cfg/virtio_fs_share_data.cfg
@@ -132,3 +132,44 @@
                     cmd_configure += 'mv config.* {0}/pjdfstest/ && mv Makefile {0}/pjdfstest/ && mv stamp-h1 {0}/pjdfstest/'
                     cmd_make = 'make %s/pjdfstest/pjdfstest'
                     cmd_pjdfstest = 'prove -rv %s/pjdfstest/tests'
+                - with_xfstest:
+                    no Windows
+                    only run_stress..with_cache.auto
+                    image_snapshot = yes
+                    mem = 32768
+                    size_mem1 = 32G
+                    blacklist = generic/551
+                    io_timeout = 14400
+                    take_regular_screendumps = no
+                    # Because screendump uses '/dev/shm' by default, it is shared with memory-backend-file.
+                    # And this case only supports Linux, and don't need to check screendump. So turn off this feature.
+                    xfstest_pkg = xfstests.tar.bz2
+                    filesystems = fs1 fs2
+                    fs_source_dir_fs1 = /tmp/virtio_fs1_test
+                    fs_source_dir_fs2 = /tmp/virtio_fs2_test
+                    fs_target_fs1 = myfs1
+                    fs_target_fs2 = myfs2
+                    fs_dest_fs1 = '/mnt/${fs_target_fs1}'
+                    fs_dest_fs2 = '/mnt/${fs_target_fs2}'
+                    cmd_unpack_xfstest = 'tar -jxvf ${fs_dest_fs1}/${xfstest_pkg} -C /home'
+                    cmd_yum_install = 'yum install -y git acl attr automake bc dump e2fsprogs fio gawk gcc libtool lvm2 '
+                    cmd_yum_install += 'make psmisc quota sed xfsdump xfsprogs libacl-devel libattr-devel libaio-devel '
+                    cmd_yum_install += 'libuuid-devel xfsprogs-devel python3 sqlite'
+                    cmd_make_xfs = 'cd /home/xfstests-dev/ && make && make install'
+                    cmd_setenv = 'echo -e 'TEST_DEV=${fs_target_fs1}\nTEST_DIR=${fs_dest_fs1}\nSCRATCH_DEV=${fs_target_fs2}\n'
+                    cmd_setenv += 'SCRATCH_MNT=${fs_dest_fs2}\nFSTYP=virtiofs\nFSX_AVOID="-E"' > configs/localhost.config'
+                    cmd_setenv += ' && echo "${blacklist}" > blacklist'
+                    cmd_xfstest = './check -virtiofs -E blacklist'
+                    cmd_useradd = 'useradd fsgqa && useradd 123456-fsgqa && useradd fsgqa2'
+                    variants:
+                        - with_local_source:
+                        - with_nfs_source:
+                            start_vm = no
+                            force_create_fs_source = no
+                            export_dir_fs1 = /tmp/virtio_fs1_test_nfs
+                            export_dir_fs2 = /tmp/virtio_fs2_test_nfs
+                            nfs_mount_src_fs1 = 127.0.0.1:/tmp/virtio_fs1_test_nfs
+                            nfs_mount_src_fs2 = 127.0.0.1:/tmp/virtio_fs2_test_nfs
+                            setup_local_nfs = yes
+                            nfs_mount_options = rw
+                            export_options = 'rw,insecure,no_root_squash,async'