New case to test virtio-fs xfstest

ID: 1906312 1907882
Signed-off-by: Zhenyu Zhang <zhenyzha@redhat.com>