@zhenyzha 2021-01-06
On aarch64:
(1/1) Host_RHEL.m8.u4.product_av.qcow2.virtio_scsi.up.virtio_net.Guest.RHEL.8.4.0.aarch64.io-github-autotest-qemu.virtio_fs_share_data.run_stress.with_xfstest.xfs.with_cache.auto.arm64-pci: FAIL: The xfstest failed. (5046.63 s)
(1/1) Host_RHEL.m8.u4.product_av.qcow2.virtio_scsi.up.virtio_net.Guest.RHEL.8.4.0.aarch64.io-github-autotest-qemu.virtio_fs_share_data.run_stress.with_xfstest.nfs.with_cache.auto.arm64-pci: FAIL: The xfstest failed. (12880.08 s)

Failure is due to product issues

@YongxueHong @menli820 pls help me to review the patch if you are free, thanks.