@zhenyzha 2021-01-07
On aarch64:
 (1/1) Host_RHEL.m8.u4.product_av.qcow2.virtio_scsi.up.virtio_net.Guest.RHEL.8.4.0.aarch64.io-github-autotest-qemu.virtio_fs_share_data.run_stress.with_xfstest.with_nfs_source.with_cache.auto.arm64-pci: PASS (16196.07 s)


@YongxueHong @menli820 pls help me to review the patch if you are free, thanks.