@menli820 2021-01-08
(3/4) Host_RHEL.m8.u4.product_av.qcow2.virtio_scsi.up.virtio_net.Guest.RHEL.8.4.0.x86_64.io-github-autotest-qemu.virtio_fs_share_data.run_stress.with_xfstest.xfs.with_cache.auto.q35: /^- FAIL: Invalid directory to remove /mnt/myfs (126540.01 s)

update the xfstest.xfs  test first, take too long time xfstest.nfs test still in running.

log:
http://10.73.72.179:8002/job-2021-01-06T10.29-4544204/test-results/3-Host_RHEL.m8.u4.product_av.qcow2.virtio_scsi.up.virtio_net.Guest.RHEL.8.4.0.x86_64.io-github-autotest-qemu.virtio_fs_share_data.run_stress.with_xfstest.xfs.with_cache.auto.q35/