@zhenyzha 2021-01-12
The following is feedback from partners and will be updated again later.

1. One is that the backend does not set 32G/dev/shm, but this should not be the root cause of the slow test run, but this will lead to OOM, if the size of the root directory is limited, now look at the size of the root directory when the beaker initializes the machine Should be enough.
2. SCRATCH_DEV and TEST_DEV should not be the same directory, usually we use TEST_DEV, TEST_DEV is mounted by default SCRATCH_DEV is equivalent to backup. Some tests require mount option, and some tests need to format the test directory. In this case, SCRATCH_DEV will be used
3. Some configurations are missing (system configuration package), which may cause some tests SKIP.
```
 useradd -m fsgqa
 groupadd fsgqa
 useradd 123456-fsgqa
 useradd fsgqa2
```