@zhenyzha 2021-01-13
> xfstests has 600+ generic tests, most of them are unavailable on virtiofs.
> About 130 tests are available to test. [RHEL8.4 xfs + virtiofs](https://docs.google.com/spreadsheets/d/12LPCe5eRcezkXyVUWlYNlT31FBiONL8Jf1ieCF3V2aE/edit#gid=364008634).
> 
> Some tests may cost a lot of time or hang on. So i think we could split virtio tests into several groups.
> Then each job will not take too long.

@liuyd96 hi 
1. Do we have a specific basis for grouping?
2. Could we accept grouping in the internal kar using test loop?