@zhenyzha 2021-01-14
> Hi, @zhenyzha
> Looks like virtiofsd option "-d" affect the test speed ( Mainly affects the reading speed )
> 
> Remove virtiofsd -o option and blacklist generic/551, the whole test costs about 1h, so i think we don't need to split tests anymore.

Okay, experiments have proved that the "-d" parameter does affect performance. I will verify it again on different platforms and update the patch later. 
Thank you very much for your assistance.@liuyd96