@liuyd96 2021-01-14
FYI, mount virtiofs and run fio tests

**Write**

cmd
```
fio --name=read_throughput --directory=$PWD --numjobs=8 --size=1G --time_based --runtime=60s --ramp_time=2s --ioengine=libaio --direct=1 --verify=0 --bs=1M --iodepth=64 --rw=write --group_reporting=1
```

without option -d
```
Run status group 0 (all jobs):
  WRITE: bw=13.8GiB/s (14.8GB/s), 13.8GiB/s-13.8GiB/s (14.8GB/s-14.8GB/s), io=829GiB (890GB), run=60001-60001msec
```

with option -d
```
Run status group 0 (all jobs):
  WRITE: bw=7956MiB/s (8343MB/s), 7956MiB/s-7956MiB/s (8343MB/s-8343MB/s), io=466GiB (501GB), run=60012-60012msec
```
 
**Read**

cmd
```
fio --name=read_throughput --directory=$PWD --numjobs=8 --size=1G --time_based --runtime=60s --ramp_time=2s --ioengine=libaio --direct=1 --verify=0 --bs=1M --iodepth=64 --rw=read --group_reporting=1
```

without option -d
```
Run status group 0 (all jobs):
   READ: bw=40.0GiB/s (42.0GB/s), 40.0GiB/s-40.0GiB/s (42.0GB/s-42.0GB/s), io=2401GiB (2578GB), run=60002-60002msec
```

with option -d
```
Run status group 0 (all jobs):
   READ: bw=14.6GiB/s (15.7GB/s), 14.6GiB/s-14.6GiB/s (15.7GB/s-15.7GB/s), io=877GiB (942GB), run=60002-60002msec
```