@liuyd96 2021-01-21
LGTM.
Tested by @liuyd96 .
xfs + virtiofs
```
[root@hpe-apollo80-02-n00 job-2021-01-20T00.45-9ba6681]# less /root/avocado/job-results/job-2021-01-20T00.45-9ba6681/job.log
[snip]
Failures: generic/003 generic/120 generic/426 generic/467 generic/477
Failed 5 of 611 tests
```
These 5 failures are known issue.